refactor: separate by classes to comply with SRP
