+++ project-e-dukate-frontend/src/components/PatientInfo/index.ts
@@ -1,0 +1,2 @@
+// src/components/PatientInfo/index.ts
+export * from './PatientInfo';
