+++ project-e-dukate-frontend/src/components/SpecialistStatusForm/index.ts
@@ -1,0 +1,2 @@
+// src/components/SpecialistStatusForm/index.ts
+export * from './SpecialistStatusForm';
