+++ project-e-dukate-frontend/src/pages/Patients/index.ts
@@ -1,2 +1,3 @@
 export { default as Patients } from './Patients';
 export { default as AddPatient } from './AddPatient';
+export { default as MedicalHistory } from './MedicalHistory';
